Revert contract working in geth fixture
